# app/static/api/openapi.yaml
openapi: 3.0.3
info:
  title: Zaviyar Perfumes & Attars POS API
  description: |
    REST API documentation for the Zaviyar Perfumes & Attars Point of Sale system.

    ## Authentication
    Most endpoints require authentication. Login via `/auth/login` to obtain a session cookie.

    ## Rate Limiting
    API requests are rate-limited to 200 per day and 50 per hour per IP address.

    ## Permissions
    Access to endpoints is controlled by role-based permissions (RBAC).
    See the permission requirements in each endpoint description.

  version: 1.0.0
  contact:
    name: API Support
    email: support@sunnatcollection.com
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/dashboard/chart-data:
    get:
      tags:
        - Dashboard
      summary: Get dashboard chart data
      description: |
        Returns data for dashboard charts including:
        - Sales trend (last 7 days)
        - Payment methods breakdown
        - Top products

        **Permission:** Authenticated user
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Chart data
          content:
            application/json:
              schema:
                type: object
                properties:
                  salesTrend:
                    type: object
                    properties:
                      labels:
                        type: array
                        items:
                          type: string
                        example: ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]
                      sales:
                        type: array
                        items:
                          type: number
                        example: [15000, 22000, 18000, 25000, 30000, 28000, 20000]
                      counts:
                        type: array
                        items:
                          type: integer
                        example: [12, 18, 15, 20, 25, 22, 16]
                  paymentMethods:
                    type: object
                    properties:
                      labels:
                        type: array
                        items:
                          type: string
                        example: ["Cash", "Card", "Online"]
                      values:
                        type: array
                        items:
                          type: number
                        example: [50000, 30000, 20000]
                  topProducts:
                    type: object
                    properties:
                      labels:
                        type: array
                        items:
                          type: string
                        example: ["Oud Perfume", "Rose Attar", "Misk"]
                      values:
                        type: array
                        items:
                          type: integer
                        example: [45, 38, 32]
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
...